[clang-doc] Add start and end line numbers
This patch adds start and end line numbers to clang-doc. Currently
clang-doc only encodes the start line numbers of records, struct, etc.
This patch adds start and end line number to clang-doc bitcode which is
passed to the generator. This will be used by the mustache backend to
generate line ranges.

Co-author: Paul Kirth <paulkirth@google.com>